Move collab/duel UI and drop task from TaskDetail to SubmitProof
The mode selector (solo/collab/duel), partner invite search, and drop
task button belong on the proof editing page, not the task viewing page.

- SubmitProof: added mode selector, invite search/dropdown/chips, drop
  task button, all with local state (no longer depends on location state)
- TaskDetail: simplified to view-only + signup button; removed collab
  state, invite functions, and mode selector UI
- TaskDetail: removed disabled "level required" button per style guide
  rule "hide controls users can't use" (Section 1.4 / Section 10)
- Sidebar: active task links now go to /tasks/{id}/submit
